# dist/tsoa/swagger.json
{
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "access_token",
                "in": "header",
                "flow": "implicit",
                "scopes": {
                    "write:pets": "modify things",
                    "read:pets": "read things"
                }
            }
        }
    },
    "info": {
        "title": ""
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/user-permission/get-all": {
            "get": {
                "operationId": "GetAllUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "User Permission"
                ],
                "security": [
                    {
                        "api_key": [
                            "read:user"
                        ]
                    }
                ],
                "parameters": []
            }
        },
        "/api/user-permission/create": {
            "post": {
                "operationId": "CreateUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "User Permission"
                ],
                "security": [
                    {
                        "api_key": [
                            "create:user"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "roles": {
                                        "items": {
                                            "type": "string"
                                        },
                                        "type": "array"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "roles",
                                    "email"
                                ],
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/api/user-permission/update/{id}": {
            "put": {
                "operationId": "UpdateUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "User Permission"
                ],
                "security": [
                    {
                        "api_key": [
                            "update:user"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "roles": {
                                        "items": {
                                            "type": "string"
                                        },
                                        "type": "array"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "roles",
                                    "email"
                                ],
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/api/user-permission/delete/{id}": {
            "delete": {
                "operationId": "DeleteUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "tags": [
                    "User Permission"
                ],
                "security": [
                    {
                        "api_key": [
                            "delete:user"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
}